+++ src/headers/arvoreAberta.h
@@ -9,7 +9,7 @@
 #include "cliente.h"
 
 typedef struct {
-    int tamanho;
+    size_t tamanho;
     Cliente **raiz;
 } ArvoreAberta;
 
@@ -17,6 +17,6 @@ ArvoreAberta *criarArvoreAberta();
 
 ArvoreAberta *preencherArvoreAberta(FILE *arquivo);
 
-void inserirAberto(ArvoreAberta *arvore, Cliente *cliente, int indice);
+void inserirAberto(ArvoreAberta *arvore, Cliente *cliente, size_t indice);
 
 #endif //TPA_ARVORE_ARVOREABERTA_H
